+++ supplemental/scripts/install-hub.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 PRODUCT_NAME="Beszel Plus"
-PRODUCT_VERSION="0.2.8"
+PRODUCT_VERSION="0.2.9"
 REPOSITORY="guzassis/beszel-plus"
 GITHUB_URL="https://github.com"
 PORT=8090